requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-filter==24.1
django-cors-headers==4.9.0
python-decouple==3.8
PyJWT==2.10.1
requests==2.32.5
